# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: TODO List API
  description: |
    A comprehensive RESTful API for managing tasks organized within multiple lists.

    ## Features
    - Create and manage multiple todo lists
    - Add, update, and delete tasks within lists
    - Set deadlines and priorities for tasks
    - Track completion status
    - Advanced querying and filtering
    - Task statistics and productivity insights

    ## Authentication
    Currently, this API does not require authentication. All endpoints are publicly accessible.

    ## Rate Limiting
    API requests are rate-limited to 1000 requests per hour per IP address.

    ## Data Validation
    All request bodies are validated against JSON schemas. Invalid requests will return 400 Bad Request with validation details.
  version: 1.0.0
  contact:
    name: Development Team
    email: dev@todoapi.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.todolist.example.com
    description: Production server
security: []
...
paths:
...
  /api/lists:
    get:
      tags: [Lists]
      summary: Get all lists
      description: Retrieve all todo lists with optional filtering and sorting
      operationId: getLists
      parameters:
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [name, createdAt, updatedAt]
          description: Field to sort by
        - in: query
          name: order
          schema:
            type: string
            enum: [asc, desc]
            default: asc
          description: Sort order
      responses:
        '200':
          description: Lists retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ListResponse'
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Lists retrieved successfully'
                  timestamp:
                    type: string
                    format: date-time
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags: [Lists]
      summary: Create a new list
      description: Create a new todo list
      operationId: createList
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateListRequest'
            example:
              name: 'Work Tasks'
              description: 'Tasks related to work projects'
      responses:
        '201':
          description: List created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ListResponse'
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'List created successfully'
                  timestamp:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/tasks:
    get:
      tags: [Queries]
      summary: Query tasks with filtering
      description: Get tasks with advanced filtering and sorting options
      operationId: queryTasks
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [completed, pending]
          description: Filter by completion status
        - in: query
          name: priority
          schema:
            type: string
            enum: [low, medium, high, urgent]
          description: Filter by priority level
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [title, priority, deadline, createdAt, updatedAt]
          description: Field to sort by
        - in: query
          name: order
          schema:
            type: string
            enum: [asc, desc]
            default: asc
          description: Sort order
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Maximum number of results
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of results to skip
      responses:
        '200':
          description: Tasks retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTaskResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
...
  schemas:
...
    ListResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the list
          example: '123e4567-e89b-12d3-a456-426614174000'
        name:
          type: string
          description: Name of the list
          example: 'Work Tasks'
        description:
          type: string
          nullable: true
          description: Optional description of the list
          example: 'Tasks related to work projects'
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the list was created
          example: '2025-07-01T10:00:00.000Z'
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the list was last updated
          example: '2025-07-02T14:30:00.000Z'
        taskCount:
          type: integer
          description: Number of tasks in the list
          example: 5
    TaskResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the task
          example: '987fcdeb-51a2-43d7-8f9e-123456789abc'
        listId:
          type: string
          format: uuid
          description: ID of the list this task belongs to
          example: '123e4567-e89b-12d3-a456-426614174000'
        title:
          type: string
          description: Title of the task
          example: 'Complete project documentation'
        description:
          type: string
          nullable: true
          description: Optional detailed description of the task
          example: 'Write comprehensive documentation for the new feature'
        completed:
          type: boolean
          description: Whether the task is completed
          example: false
        priority:
          type: string
          enum: [low, medium, high, urgent]
          description: Priority level of the task
          example: 'high'
        deadline:
          type: string
          format: date-time
          nullable: true
          description: Optional deadline for the task
          example: '2025-07-10T17:00:00.000Z'
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the task was created
          example: '2025-07-01T11:00:00.000Z'
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the task was last updated
          example: '2025-07-02T15:00:00.000Z'
        completedAt:
          type: string
          format: date-time
          nullable: true
          description: Timestamp when the task was completed
          example: null

    # Request Models
    CreateListRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Name of the list
          example: 'Work Tasks'
        description:
          type: string
          maxLength: 500
          description: Optional description of the list
          example: 'Tasks related to work projects'
...
    PaginatedTaskResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/TaskResponse'
        pagination:
          type: object
          properties:
            total:
              type: integer
              description: Total number of items
              example: 100
            limit:
              type: integer
              description: Number of items per page
              example: 20
            offset:
              type: integer
              description: Number of items skipped
              example: 0
            hasNext:
              type: boolean
              description: Whether there are more items
              example: true
            hasPrev:
              type: boolean
              description: Whether there are previous items
              example: false
        success:
          type: boolean
          example: true
        message:
          type: string
          example: 'Tasks retrieved successfully'
        timestamp:
          type: string
          format: date-time

    # Error Models
    ValidationError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: 'VALIDATION_ERROR'
            message:
              type: string
              example: 'Validation failed'
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                    example: 'title'
                  message:
                    type: string
                    example: 'Title is required'
                  value:
                    oneOf:
                      - type: string
                      - type: number
                      - type: boolean
                      - type: 'null'
        timestamp:
          type: string
          format: date-time
...
    InternalServerError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: 'INTERNAL_SERVER_ERROR'
            message:
              type: string
              example: 'An unexpected error occurred'
            requestId:
              type: string
              example: 'req_123456789'
        timestamp:
          type: string
          format: date-time
  responses:
    ValidationError:
      description: Bad Request - Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
...
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternalServerError'
...